Adds deploy workflow with GitHub Pages
Sets up a GitHub Actions workflow to automatically deploy the application to GitHub Pages upon pushes to the main branch.

This includes installing dependencies, building the project, and deploying the contents of the dist directory to the `gh-pages` branch.
Also adds `gh-pages` as a dev dependency and a deploy script in `package.json` for manual deployment.

Sets base URL for Vite to correctly serve the application from the GitHub Pages subdirectory.